+++ examples/verilog-project/tests/dpi_tutorial.rs
@@ -12,11 +12,9 @@
 // You should have received a copy of the GNU General Public License along with
 // this program.  If not, see <https://www.gnu.org/licenses/>.
 
-use std::env;
-
 use example_verilog_project::{DpiMain, MoreDpiMain};
 use marlin::{
-    verilator::{VerilatorRuntime, VerilatorRuntimeOptions},
+    verilator::{VerilatorRuntime, VerilatorRuntimeOptions, verilator_version},
     verilog::prelude::*,
 };
 use snafu::Whatever;
@@ -31,16 +29,13 @@ pub extern "C" fn set_out(output: &mut i32) {
 #[test]
 #[snafu::report]
 fn main_tutorial() -> Result<(), Whatever> {
-    if env::var("RUST_LOG").is_ok() {
-        env_logger::init();
-    }
-
     let runtime = VerilatorRuntime::new(
         "artifacts".into(),
         &["src/dpi.sv".as_ref()],
         &[],
         [set_out],
-        VerilatorRuntimeOptions::default_logging(),
+        VerilatorRuntimeOptions::default()
+            .allow_unsupported_verilator(Some(verilator_version!(5 020))),
     )?;
 
     let mut main = runtime.create_model_simple::<DpiMain>()?;
@@ -77,7 +72,8 @@ fn other_test() -> Result<(), Whatever> {
         &["src/more_dpi.sv".as_ref()],
         &[],
         [set_unsigned_int_out, check_unsigned_int_out, set_bool_out],
-        VerilatorRuntimeOptions::default_logging(),
+        VerilatorRuntimeOptions::default()
+            .allow_unsupported_verilator(Some(verilator_version!(5 020))),
     )?;
 
     let mut main = runtime.create_model_simple::<MoreDpiMain>()?;