+++ examples/verilog-project/tests/tracing.rs
@@ -16,6 +16,7 @@ use example_verilog_project::Main;
 use marlin::{
     verilator::{
         VerilatedModelConfig, VerilatorRuntime, VerilatorRuntimeOptions,
+        tracing::Waveform, verilator_version,
     },
     verilog::prelude::*,
 };
@@ -23,21 +24,21 @@ use snafu::Whatever;
 
 #[test]
 #[snafu::report]
-fn forwards_correctly() -> Result<(), Whatever> {
+fn forwards_correctly_vcd() -> Result<(), Whatever> {
     let runtime = VerilatorRuntime::new(
         "artifacts".into(),
         &["src/main.sv".as_ref()],
         &[],
         [],
-        VerilatorRuntimeOptions::default_logging(),
+        VerilatorRuntimeOptions::default()
+            .allow_unsupported_verilator(Some(verilator_version!(5 020))),
     )?;
 
-    let mut main = runtime.create_model::<Main>(&VerilatedModelConfig {
-        enable_tracing: true,
-        ..Default::default()
-    })?;
+    let mut main = runtime.create_model::<Main>(
+        &VerilatedModelConfig::default().enable_tracing(Some(Waveform::Vcd)),
+    )?;
 
-    let mut vcd = main.open_vcd("foo.vcd");
+    let mut vcd = main.open_trace("foo.vcd");
 
     vcd.dump(0);
 
@@ -50,6 +51,39 @@ fn forwards_correctly() -> Result<(), Whatever> {
 
     vcd.dump(1);
     vcd.dump(2);
+
+    Ok(())
+}
+
+#[test]
+#[snafu::report]
+fn forwards_correctly_fst() -> Result<(), Whatever> {
+    let runtime = VerilatorRuntime::new(
+        "artifacts".into(),
+        &["src/main.sv".as_ref()],
+        &[],
+        [],
+        VerilatorRuntimeOptions::default()
+            .allow_unsupported_verilator(Some(verilator_version!(5 020))),
+    )?;
+
+    let mut main = runtime.create_model::<Main>(
+        &VerilatedModelConfig::default().enable_tracing(Some(Waveform::Fst)),
+    )?;
+
+    let mut fst = main.open_trace("foo.fst");
+
+    fst.dump(0);
+
+    main.medium_input = u32::MAX;
+    println!("{}", main.medium_output);
+    assert_eq!(main.medium_output, 0);
+    main.eval();
+    println!("{}", main.medium_output);
+    assert_eq!(main.medium_output, u32::MAX);
+
+    fst.dump(1);
+    fst.dump(2);
 
     Ok(())
 }