+++ language-support/verilog/src/lib.rs
@@ -16,6 +16,8 @@ pub use marlin_verilog_macro::dpi;
 
 pub mod prelude {
     pub use crate as verilog;
-    pub use marlin_verilator::{AsDynamicVerilatedModel, AsVerilatedModel};
+    pub use marlin_verilator::{
+        AsDynamicVerilatedModel, AsVerilatedModel, tracing::OpenTrace,
+    };
     pub use marlin_verilog_macro::verilog;
 }